requirements.txt
asgiref==3.5.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.6.0
firebase-admin==5.2.0
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.1
pipenv==2022.4.21
platformdirs==2.5.2
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
